Move UDF into utils
