+++ fonduer/parser/parser.py
@@ -21,7 +21,7 @@ from fonduer.models import (
     split_stable_id,
 )
 from fonduer.parser.spacy_parser import Spacy
-from fonduer.udf import UDF, UDFRunner
+from fonduer.utils.udf import UDF, UDFRunner
 from fonduer.parser.visual import VisualLinker
 
 logger = logging.getLogger(__name__)
